requirements.txt
absl-py==1.4.0
accelerate==0.20.3
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array-record==0.2.0
arrow==1.3.0
art==6.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.2
biosignalsnotebooks==0.6.13
bleach==6.1.0
bokeh==2.4.3
bs4==0.0.2
cachetools==5.3.2
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cssselect2==0.7.0
cycler==0.12.1
dask==2024.1.0
DateTime==5.4
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distributed==2024.1.0
dm-tree==0.1.8
dnspython==2.6.1
EasyProcess==1.1
einops==0.7.0
entrypoint2==1.1
et-xmlfile==1.1.0
etils==1.2.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.9.0
filterpy==1.4.5
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
glasses @ git+https://github.com/FrancescoSaverioZuppichini/glasses@b41cbe2a49e76fccf05852c4d365c554e2cb9d5f
google-auth==2.26.2
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.59.0
grpcio==1.60.0
h5py==3.10.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
importlib-resources==5.12.0
ipykernel==6.28.0
ipython==8.20.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.4
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kaggle==1.6.3
kiwisolver==1.4.5
libmagic==1.0
lightning==2.1.3
lightning-utilities==0.10.0
llvmlite==0.41.1
locket==1.0.0
lxml==5.2.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.7
mss==9.0.1
multidict==6.0.4
nbclient==0.9.0
nbconvert==7.14.1
nbformat==5.9.2
nest-asyncio==1.5.9
networkx==3.0
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opendatasets==0.1.22
openpyxl==3.1.2
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.1
patsy==0.5.6
Pillow==9.3.0
platformdirs==4.1.0
prometheus-client==0.19.0
promise==2.3
prompt-toolkit==3.0.43
protobuf==4.23.0
psutil==5.9.7
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycm==4.0
pycparser==2.21
pydub==0.25.1
pygame==2.4.0
PyGetWindow==0.0.9
Pygments==2.17.2
pylsl==1.16.2
pymongo==4.7.2
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
python-json-logger==2.0.7
python-magic-bin==0.4.14
python-slugify==8.0.1
python-vlc==3.0.20123
pytorch-accelerated==0.1.47
pytorch-lightning==2.1.3
pyts==0.13.0
pytweening==1.0.7
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.8.1
seaborn==0.13.2
Send2Trash==1.8.2
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.1
stumpy==1.12.0
sympy==1.12
tblib==3.0.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow-metadata==1.13.1
termcolor==2.3.0
terminado==0.18.0
text-unidecode==1.3
threadpoolctl==3.2.0
timm==0.9.12
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.1
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchinfo==1.8.0
torchmetrics==1.3.0
torchvision==0.16.2+cu121
tornado==6.4
tqdm==4.64.1
traitlets==5.14.1
tsfel==0.1.6
tsfresh==0.20.1
tslearn==0.6.3
tushare==1.4.6
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
uri-template==1.3.0
urllib3==2.1.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==3.0.1
wget==3.2
xlwt==1.3.0
yarl==1.9.4
zict==3.0.0
zipp==3.17.0
zope.interface==6.1
